requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
APScheduler==3.11.2
hdbcli==2.28.17
httpx==0.28.1
Jinja2==3.1.6
python-dotenv==1.2.2
tenacity==9.1.4
slowapi==0.1.9
prometheus-client==0.25.0
